Add tests to testapi for null OpaqueJSStrings https://bugs.webkit.org/show_bug.cgi?id=98805
Reviewed by Geoffrey Garen.

Added tests that check that OpaqueJSString, which is wrapped via JSStringRef, properly returns
null strings and that a null string in a JSStringRef will return a NULL JSChar* and 0 length
via the JSStringGetCharactersPtr() and JSStringGetLength() APIs respectively. Added a check that
JSValueMakeFromJSONString() properly handles a null string as well.

* API/tests/testapi.c:
(main):


Identifier: 116816@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@130832 268f45cc-cd09-0410-ab3c-d52691b4dbfc