Needs internal API to return distributed nodes for InsertionPoint https://bugs.webkit.org/show_bug.cgi?id=98868
Reviewed by Hajime Morita.

.:

* Source/autotools/symbols.filter:

Source/WebCore:

When testing insertion points, we would like to have an internal API which returns a node list
distributed to an insertion point.

We've added the API in Internals.

Test: fast/dom/shadow/distributed-nodes.html

* WebCore.exp.in:
* html/shadow/InsertionPoint.cpp:
(WebCore::InsertionPoint::distributedNodes): Returns the distributed nodes. When InsertionPoint is not
attached, the result will be null, since the distribution is not calculated correctly.
(WebCore):
* html/shadow/InsertionPoint.h:
(InsertionPoint):
* testing/Internals.cpp:
(WebCore::Internals::distributedNodes):
(WebCore):
* testing/Internals.h:
(Internals):
* testing/Internals.idl:

Source/WebKit2:

* win/WebKit2.def:
* win/WebKit2CFLite.def:

LayoutTests:

* fast/dom/shadow/distributed-nodes-expected.txt: Added.
* fast/dom/shadow/distributed-nodes.html: Added.


Identifier: 116903@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@130926 268f45cc-cd09-0410-ab3c-d52691b4dbfc