[EFL] More unreviewed gardening.
Revert my r130959, it turns out the tests being reported as
failing were ones which were not listed in TestExpectations at
all, so no cookie for me.

Also add `Missing' as a valid expectation for
fast/css/text-rendering.html. While at it remove the text
expectations for it, since the test is failing for us anyway.

* platform/efl/TestExpectations:
* platform/efl/fast/css/text-rendering-expected.txt: Removed.


Identifier: 117008@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@131038 268f45cc-cd09-0410-ab3c-d52691b4dbfc