[EFL] Add background color to the radio button theme. https://bugs.webkit.org/show_bug.cgi?id=98852
Patch by Byungwoo Lee <bw80.lee@samsung.com> on 2012-10-17
Reviewed by Gyuyoung Kim.

Background color is added to the radio button theme for identifying
radio button status easily.

Radio button theme uses images which has transparent background.

When the background of radio button is black, checked/unchecked status
cannot be distinguished because check mark color is black.

And it is hard to identify whether the radio button is disabled or not,
because there is no background color in the radio button theme.

* DefaultTheme/widget/radio/img_radio_bg_disabled.png: Added.
* DefaultTheme/widget/radio/img_radio_bg_enabled.png: Added.
* DefaultTheme/widget/radio/radio.edc:

Identifier: 117511@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@131554 268f45cc-cd09-0410-ab3c-d52691b4dbfc