fast/forms/range/input-appearance-range-rtl.html off by one pixel https://bugs.webkit.org/show_bug.cgi?id=99625
Reviewed by Ojan Vafai.

Source/WebCore:

Previously, we would assume that a vertical slider in RTL would render identically to
a vertical slider in LTR. Due to differing thumb sizes, there is an off by one.

This worked in deprecated flexbox because it doesn't adjust for RTL when applying
box-align: center.

Tests: fast/forms/range/input-appearance-range-rtl.html

* html/shadow/SliderThumbElement.cpp:
(WebCore::RenderSliderContainer::layout): Force LTR for flexbox layout when vertical. We have to restore the old value
just in case the user switches from vertical to horizontal.

LayoutTests:

Run test on GTK+ and Qt now that it is fixed.

* platform/gtk/TestExpectations:
* platform/qt/TestExpectations:


Identifier: 117612@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@131657 268f45cc-cd09-0410-ab3c-d52691b4dbfc