Incorrect rendering of borders on <col> with span > 1 https://bugs.webkit.org/show_bug.cgi?id=76246
Patch by Arpita Bahuguna <arpitabahuguna@gmail.com> on 2012-10-17
Reviewed by Julien Chaffraix.

Source/WebCore:

The HTML5 rendering specification [10.2.2 - Display Types] states that
"For the purposes of the CSS table model, the col element is expected to
be treated as if it was present as many times as its span attribute
specifies."
We should thus apply a col element's border as if the element is present
as many number of times as its span attribute.

Apart from this, we should also treat the col and its enclosing colgroup
separately while computing the collapsed borders.

Test: fast/table/border-collapsing/collapsed-border-with-col-colgroup-span.html

* rendering/RenderTableCell.cpp:
(WebCore::RenderTableCell::computeCollapsedStartBorder):
(WebCore::RenderTableCell::computeCollapsedEndBorder):
Borders from col and its enclosing colgroup element should be handled
separately, while considering the preceeding col's end border (for start
border computation) and the next col's start border (for end border
computation).

Also, have made changes for handling of col elements with span attribute as
per the specification. We now apply the border (start or end) of the col
element irrespective of whether it has any span specified for it or not.

LayoutTests:

* fast/table/border-collapsing/cached-change-colgroup-border-color-expected.png:
* fast/table/border-collapsing/cached-change-colgroup-border-color-expected.txt:
* fast/table/border-collapsing/cached-change-colgroup-border-width-expected.png:
* fast/table/border-collapsing/cached-change-colgroup-border-width-expected.txt:
Existing tests modified. This is because previously, while computing collapsed
start border, we did not take the preceeding col's enclosing colgroup's end border
into consideration. While computing the collapsed start border, only the preceeding
col element's end border was considered.

With this fix, for the above two tests, the last col's width now changes due to
the border being applied to it (the preceeding col's enclosing colgroup's end border)
which causes the table's width to change. Also, we should note that the cell's grow
by half the border's width; which is expected (on account of their being collapsed
borders).

* fast/table/border-collapsing/collapsed-border-with-col-colgroup-span-expected.txt: Added.
* fast/table/border-collapsing/collapsed-border-with-col-colgroup-span.html: Added.
* platform/chromium-linux/fast/table/border-collapsing/collapsed-border-with-col-colgroup-span-expected.png: Added.
Image only test added for verifying the behavior of collapsed borders with
col and colgroup span.
New expected image file added for the chromium-linux port.

Identifier: 117625@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@131671 268f45cc-cd09-0410-ab3c-d52691b4dbfc