StyleRareNonInheritedData::contentDataEquivalent only looks at the first ContentData https://bugs.webkit.org/show_bug.cgi?id=99560
Patch by Elliott Sprehn <esprehn@chromium.org> on 2012-10-17
Reviewed by Eric Seidel.

Source/WebCore:

Previously we only compared the first ContentData in the linked
list of ContentData's which meant that if the resolved style
for content had the same prefix we wouldn't update the content.

This patch adds a loop to compare each of the ContentData objects
in the linked list.

Test: fast/css-generated-content/content-property-change.html

* rendering/style/StyleRareNonInheritedData.cpp:
(WebCore::StyleRareNonInheritedData::contentDataEquivalent):

LayoutTests:

Test that when the content CSS property changes we update the displayed content.

* fast/css-generated-content/content-property-change-expected.html: Added.
* fast/css-generated-content/content-property-change.html: Added.

Identifier: 117639@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@131685 268f45cc-cd09-0410-ab3c-d52691b4dbfc